requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
asyncpraw
pydantic
pydantic-settings
python-dotenv
openai
vaderSentiment
weaviate-client
httpx
backoff
loguru
google-api-python-client
streamlit
pandas
plotly
pytest
pytest-asyncio
mypy
python-dateutil
